AUTH: Support basic auth for local ai
By default localai.io does not come with any authentication.

It is however trival to run behind a reverse proxy, and the most
common way to authenticate with any webserver is with the widely
supported basic auth scheme, using a username and password.

This patch allows an administrator or user to set a username and
password to use with basic auth for a local installation of localai.

Signed-off-by: Micke Nordin <kano@sunet.se>